requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu